51 refactored ScenarioOutlineRunner.GetString to GetDisplayName
