+++ src/Xbehave.2.Execution/Xbehave.2.Execution.csproj
@@ -56,6 +56,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Guard.cs" />
     <Compile Include="ScenarioOutline.cs" />
+    <Compile Include="TypeInfoExtensions.cs" />
     <Compile Include="ScenarioOutlineRunner.cs" />
     <Compile Include="ScenarioRunner.cs" />
     <Compile Include="Shims\ExceptionExtensions.cs" />
